Initial commit for 1.20.1, mostly works but still needs bug fixing

+++ northstar/src/main/java/com/lightning/northstar/particle/ColdAirParticle.java
@@ -52,7 +52,7 @@ public class ColdAirParticle  extends SimpleAnimatedParticle {
 
 	@Override
 	public int getLightColor(float partialTick) {
-		BlockPos blockpos = new BlockPos(this.x, this.y, this.z);
+		BlockPos blockpos = new BlockPos((int)this.x, (int)this.y, (int)this.z);
 		return this.level.isLoaded(blockpos) ? LevelRenderer.getLightColor(level, blockpos) : 0;
 	}
